@@ -6,7 +6,7 @@
   "types": "./lib/index.d.ts",
   "scripts": {
     "download-x": "cd x2JavaScript && ./generate && cd ..",
-    "generate": "node post-process.js && npm run generate-all-types && npm run format",
+    "generate": "npx ts-node post-process.ts && npm run generate-all-types && npm run format",
     "generate-all-types": "DESTINATION=../src/xdr MAIN_FILE_NAME=types.rs npx ts-node x2JavaScript/generated/stellar-xdr_generated.ts",
     "format": "rustfmt ../src/xdr/types.rs",
     "build": "npm run download-x && npm run generate"
@@ -16,8 +16,8 @@
   "devDependencies": {
     "@types/node": "^15.12.2",
     "js-xdr": "file:./js-xdr",
-    "ts-node": "^10.0.0",
-    "typescript": "^4.3.2"
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.5"
   },
   "repository": {
     "type": "git",
