@@ -1,6 +1,6 @@
 //! Autogenerated XDR types
 //!
-// This code has been automatically generated on 2023-10-02
+// This code has been automatically generated on 2023-10-05
 // using the code in folder `/autogenerator`
 // Do not edit this file manually!
 
